Crate zng_ext_fs_watcher

File system events and service.

Crate

This crate is part of the zng project.

Cargo Features

This crate provides 4 feature flags, 0 enabled by default.

"json"

Enable JSON helpers.

"yaml"

Enable YAML helpers.

"toml"

Enable TOML helpers.

"ron"

Enable RON helpers.

Modules

fs_event

File system change event types.

Structs

FsChange

Represents a single file system change, annotated.

FsChangeNoteHandle

Handle that holds a WATCHER.annotate note.

FsChangesArgs

FS_CHANGES_EVENT arguments.

FsWatcherManager

Application extension that provides file system change events and service.

WATCHER

File system watcher service.

WatchFile

Represents an open read-only file provided by WATCHER.read.

WatcherHandle

Represents an active file or directory watcher in WATCHER.

WatcherSyncWriteNote

Annotation for file watcher events and var update tags.

WriteFile

Represents an open write file provided by WATCHER.sync.

Enums

WatchFileParseError

Error for WatchFile::parse.

Statics

FS_CHANGES_EVENT

Event sent by the WATCHER service on directories or files that are watched.

Traits

FsChangeNote

Represents a FsChange note.

WatcherReadStatus

Represents a status type for WATCHER read-only operations.

WatcherSyncStatus

Represents a status type for WATCHER.sync_status.