zng_ext_fs_watcher
Module fs_event

File system change event types.

The event for each change is available in FsChange::event.

This module re-exports types from the notify crate.

Structs

Error

Notify error type.

Event

Notify event.

Enums

AccessKind

An event describing non-mutating access operations on files.

AccessMode

An event describing open or close operations on files.

CreateKind

An event describing creation operations on files.

DataChange

An event emitted when the data content of a file is changed.

ErrorKind

Error kinds

EventKind

Top-level event kind.

MetadataKind

An event emitted when the metadata of a file or folder is changed.

ModifyKind

An event describing mutation of content, name, or metadata.

RemoveKind

An event describing removal operations on files.

RenameMode

An event emitted when the name of a file or folder is changed.